+++ nodalia-cards-2.0.0.js
@@ -1,2 +1,2 @@
-import "./nodalia-cards-2.0.1.js";
-if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.0.1","contentSha256_12":"a65962216ea6","file":"nodalia-cards-2.0.0.js","targetFile":"nodalia-cards-2.0.1.js","fallbackFile":"nodalia-cards.js"};}
+import "./nodalia-cards-2.0.2.js";
+if(typeof window!=="undefined"){window.__NODALIA_LOADER__={"mode":"compat","pkgVersion":"2.0.2","contentSha256_12":"ca3834a74bc0","file":"nodalia-cards-2.0.0.js","targetFile":"nodalia-cards-2.0.2.js","fallbackFile":"nodalia-cards.js"};}