+++ nodalia-cards.manifest.js
@@ -1,23 +1,23 @@
 export default {
-  "pkgVersion": "2.0.1",
-  "contentSha256_12": "a65962216ea6",
+  "pkgVersion": "2.0.2",
+  "contentSha256_12": "ca3834a74bc0",
   "file": "nodalia-cards.bundle.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "compatLoaderFiles": [
-    "nodalia-cards-2.0.0-alpha.60.js",
-    "nodalia-cards-2.0.0.js"
+    "nodalia-cards-2.0.0.js",
+    "nodalia-cards-2.0.1.js"
   ],
-  "splitCoreFile": "nodalia-cards-core-2.0.1.js",
-  "splitCoreSha256_12": "ddd1431fcb0d",
-  "splitSuiteFile": "nodalia-cards-suite-2.0.1.js",
-  "splitSuiteSha256_12": "befca8aca11c",
-  "editorFile": "nodalia-cards-editor-2.0.1.js",
+  "splitCoreFile": "nodalia-cards-core-2.0.2.js",
+  "splitCoreSha256_12": "b725c8b161c6",
+  "splitSuiteFile": "nodalia-cards-suite-2.0.2.js",
+  "splitSuiteSha256_12": "b3582d52485d",
+  "editorFile": "nodalia-cards-editor-2.0.2.js",
   "editorSha256_12": "cf952c17c474"
 };
-export const pkgVersion = "2.0.1";
-export const contentSha256_12 = "a65962216ea6";
+export const pkgVersion = "2.0.2";
+export const contentSha256_12 = "ca3834a74bc0";
 export const file = "nodalia-cards.bundle.js";
-export const splitCoreFile = "nodalia-cards-core-2.0.1.js";
-export const splitSuiteFile = "nodalia-cards-suite-2.0.1.js";
-export const editorFile = "nodalia-cards-editor-2.0.1.js";
+export const splitCoreFile = "nodalia-cards-core-2.0.2.js";
+export const splitSuiteFile = "nodalia-cards-suite-2.0.2.js";
+export const editorFile = "nodalia-cards-editor-2.0.2.js";